removed all orig files
added another network test
