+++ package/tests/test_services/test_services.py
@@ -1,17 +1,15 @@
 from unittest import TestCase
 
 import mock
-from azure.mgmt.network.models import IPAllocationMethod
 from azure.mgmt.storage.models import StorageAccountCreateParameters
 from mock import MagicMock
 from mock import Mock
 from msrestazure.azure_operation import AzureOperationPoller
 
-from cloudshell.cp.azure.domain.services.network_service import NetworkService
 from cloudshell.cp.azure.domain.services.key_pair import KeyPairService
+from cloudshell.cp.azure.domain.services.security_group import SecurityGroupService
 from cloudshell.cp.azure.domain.services.storage_service import StorageService
 from cloudshell.cp.azure.domain.services.virtual_machine_service import VirtualMachineService
-from cloudshell.cp.azure.domain.services.security_group import SecurityGroupService
 from tests.helpers.test_helper import TestHelper
 
 
@@ -90,71 +88,6 @@ class TestStorageService(TestCase):
             storage_name=self.storage_name)
 
         self.assertEqual(key, storage_key.value)
-
-
-class TestNetworkService(TestCase):
-    def setUp(self):
-        self.network_service = NetworkService()
-
-    def test_network_for_vm_fails_when_public_ip_type_is_not_correct(self):
-
-        self.assertRaises(Exception,
-                          self.network_service.create_network_for_vm,
-                          network_client=MagicMock(),
-                          group_name=Mock(),
-                          interface_name=Mock(),
-                          ip_name=Mock(),
-                          region=Mock(),
-                          subnet=Mock(),
-                          tags=Mock(),
-                          add_public_ip=True,
-                          public_ip_type="a_cat")
-
-    def test_vm_created_with_private_ip_static(self):
-        # Arrange
-
-        region = "us"
-        management_group_name = "company"
-        interface_name = "interface"
-        network_name = "network"
-        subnet_name = "subnet"
-        ip_name = "ip"
-        tags = "tags"
-
-        network_client = MagicMock()
-        network_client.virtual_networks.create_or_update = MagicMock()
-        network_client.subnets.get = MagicMock()
-        network_client.public_ip_addresses.create_or_update = MagicMock()
-        network_client.public_ip_addresses.get = MagicMock()
-        result = MagicMock()
-        result.result().ip_configurations = [MagicMock()]
-        network_client.network_interfaces.create_or_update = MagicMock(return_value=result)
-
-        # Act
-        self.network_service.create_network_for_vm(
-            network_client=network_client,
-            group_name=management_group_name,
-            interface_name=interface_name,
-            ip_name=ip_name,
-            region=region,
-            subnet=MagicMock(),
-            add_public_ip=True,
-            public_ip_type="Static",
-            tags=tags)
-
-        # Verify
-
-        self.assertEqual(network_client.network_interfaces.create_or_update.call_count, 2)
-
-        # first time dynamic
-        self.assertEqual(network_client.network_interfaces.create_or_update.call_args_list[0][0][2].ip_configurations[
-                             0].private_ip_allocation_method,
-                         IPAllocationMethod.dynamic)
-
-        # second time static
-        self.assertEqual(network_client.network_interfaces.create_or_update.call_args_list[1][0][2].ip_configurations[
-                             0].private_ip_allocation_method,
-                         IPAllocationMethod.static)
 
 
 class TestVMService(TestCase):
